refactoring - makepot core
